+++ NEWS.md
@@ -1,4 +1,5 @@
 ## Version `v3.1.0-SNAPSHOT` (TBD)
+* Adopt APPPOCTOOL-85 Kafka producer module split by moving from `folio-integration-kafka` to `folio-kafka-producer` (APPPOCTOOL-85)
 * Fix parameter names and typos in README curl examples (MODLOGINKC-51)
 * User can't log in due to invalid token error (MODLOGINKC-44)
 * Introduce configuration for FSSP (APPPOCTOOL-59)